Move `SelfAssignment` into `Operators` lint pass
