+++ clippy_lints/src/lib.register_lints.rs
@@ -437,6 +437,7 @@ store.register_lints(&[
     operators::NEEDLESS_BITWISE_BOOL,
     operators::OP_REF,
     operators::PTR_EQ,
+    operators::SELF_ASSIGNMENT,
     operators::VERBOSE_BIT_MASK,
     option_env_unwrap::OPTION_ENV_UNWRAP,
     option_if_let_else::OPTION_IF_LET_ELSE,
@@ -483,7 +484,6 @@ store.register_lints(&[
     returns::LET_AND_RETURN,
     returns::NEEDLESS_RETURN,
     same_name_method::SAME_NAME_METHOD,
-    self_assignment::SELF_ASSIGNMENT,
     self_named_constructors::SELF_NAMED_CONSTRUCTORS,
     semicolon_if_nothing_returned::SEMICOLON_IF_NOTHING_RETURNED,
     serde_api::SERDE_API_MISUSE,
